+++ src/main/java/com/sparta/deliveryminiproject/domain/cart/dto/CartRequestDto.java
@@ -1,6 +1,6 @@
-package com.sparta.deliveryminiproject.domain.order.dto;
+package com.sparta.deliveryminiproject.domain.cart.dto;
 
-import com.sparta.deliveryminiproject.domain.order.entity.Cart;
+import com.sparta.deliveryminiproject.domain.cart.entity.Cart;
 import com.sparta.deliveryminiproject.domain.shop.entity.Menu;
 import com.sparta.deliveryminiproject.domain.user.entity.User;
 import jakarta.validation.constraints.NotBlank;
