+++ src/main/java/com/sparta/deliveryminiproject/domain/cart/entity/Cart.java
@@ -1,5 +1,6 @@
-package com.sparta.deliveryminiproject.domain.order.entity;
+package com.sparta.deliveryminiproject.domain.cart.entity;
 
+import com.sparta.deliveryminiproject.domain.order.entity.Order;
 import com.sparta.deliveryminiproject.domain.shop.entity.Menu;
 import com.sparta.deliveryminiproject.domain.shop.entity.Shop;
 import com.sparta.deliveryminiproject.domain.user.entity.User;
