+++ src/test/java/com/sparta/deliveryminiproject/domain/order/service/CartServiceTest.java
@@ -3,9 +3,10 @@ package com.sparta.deliveryminiproject.domain.order.service;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.sparta.deliveryminiproject.domain.order.dto.CartRequestDto;
-import com.sparta.deliveryminiproject.domain.order.entity.Cart;
-import com.sparta.deliveryminiproject.domain.order.repository.CartRepository;
+import com.sparta.deliveryminiproject.domain.cart.dto.CartRequestDto;
+import com.sparta.deliveryminiproject.domain.cart.entity.Cart;
+import com.sparta.deliveryminiproject.domain.cart.repository.CartRepository;
+import com.sparta.deliveryminiproject.domain.cart.service.CartService;
 import com.sparta.deliveryminiproject.domain.shop.entity.Menu;
 import com.sparta.deliveryminiproject.domain.shop.entity.Shop;
 import com.sparta.deliveryminiproject.domain.shop.repository.MenuRepository;
